services/users/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["vedaka1 <vladik.kuleshov.2003@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.0"
sqlalchemy = "^2.0.31"
asyncpg = "^0.29.0"
dishka = "^1.2.0"
pyjwt = {extras = ["crypto"], version = "^2.9.0"}
aiohttp = "^3.10.0"
uvicorn = "^0.30.5"


[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
